here is

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	float d;
 	long long int e;
 
-	printf("Size of a char: %lu byte(s)\n",sizeof(c));
-	printf("Size of an int: %lu byte(s)\n",sizeof(a));
-	printf("Size of a long int: %lu byte(s)\n",sizeof(b));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(e));
-	printf("Size of a float: %lu byte(s)\n",sizeof(d));
+	printf("Size of a char: %lu byte(s)\n",(unsigned long)sizeof(c));
+	printf("Size of an int: %lu byte(s)\n",(unsigned long)sizeof(a));
+	printf("Size of a long int: %lu byte(s)\n",(unsigned long)sizeof(b));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(e));
+	printf("Size of a float: %lu byte(s)\n",(unsigned long)sizeof(d));
         return (0);
 }
